pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jntts-cli"  
version = "1.1.0" 
authors = [
  { name="Justin Nguyen 🇻🇳", email="duchuy_1997@hotmail.com" },
]
description = "Một công cụ TTS sử dụng mô hình Bark/Suno để chuyển đổi văn bản thành giọng nói."
readme = "README.md"  
requires-python = ">=3.9"
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: Multimedia :: Sound/Audio :: Speech"
]

# PHẦN QUAN TRỌNG NHẤT: Dependencies
dependencies = [
    "torch",
    "transformers",
    "sounddevice",
    "tqdm",
    "pydub",
    "pyfiglet",
    "psutil",
    "pynvml",
    "py-cpuinfo",
    "appdirs",
    "scipy",
    "importlib-resources",
    "accelerate>=0.33.0",
    "bitsandbytes>0.37.0; platform_machine != 'arm64' and platform_system != 'Darwin'",
    "cached_path",
    "click",
    "datasets",
    "ema_pytorch>=0.5.2",
    "gradio>=3.45.2",
    "hydra-core>=1.3.0",
    "jieba",
    "librosa",
    "matplotlib",
    "numpy<=1.26.4",
    "pypinyin",
    "safetensors",
    "soundfile",
    "tomli",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "torchdiffeq",
    "tqdm>=4.65.0",
    "transformers_stream_generator",
    "vocos",
    "wandb",
    "x_transformers>=1.31.14",
    "openai-whisper",
    "ffmpeg",
    "f5-tts",
    "wcwidth"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"] 
include = ["jntts*"]  

[project.scripts]
jntts-cli = "jntts.main:main"